现在职业可以发放物品了(死亡不掉落) 现在可以禁用职业/进度奖励了

+++ src/main/java/net/mcbbs/jerryzrf/minehunt/util/StatisticsBaker.java
@@ -52,7 +52,7 @@ public class StatisticsBaker {
                 continue;
             }
             maxPlayer = playerDoubleEntry.getKey();
-            maxDamage = playerDoubleEntry.getValue().intValue();
+            maxDamage = playerDoubleEntry.getValue().doubleValue();
         }
         if (maxPlayer == null) {
             return "";
